화면왜곡

+++ KatanaZero/MainGame.h
@@ -34,6 +34,6 @@ private:
 	Timer* timer;
 
 	GPImage testDraw;
-	int tmpTimer;
+	float tmpTimer;
 };
